Consistent Hashing
libconhash is a consistent hashing library which can be compiled both on
Windows and Linux platforms, with the following features:

High performance and easy to use, libconhash uses a red-black tree to
manage all nodes to achieve high performance.
By default, it uses the MD5 algorithm, but it also supports user-defined
hash functions.
Easy to scale according to the node's processing capacity.